requirements.txt
flask
waitress
python-whois